+++ CHANGELOG.md
@@ -5,7 +5,7 @@
     Click to see more.
   </summary>
 
-  - Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
+  - Minor: Create Listbox component
 </details>
 
 ## 0.23.0 (June 11, 2020)