+++ CHANGELOG.md
@@ -5,10 +5,11 @@
     Click to see more.
   </summary>
 
-  - Minor: Add linter for Typography
-
   - Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
 </details>
+
+## 0.24.0 (June 16, 2020)
+  - Minor: Add linter for Typography
 
 ## 0.23.0 (June 11, 2020)
   - Major: remove unnecessary styling (inline-block and vertical alignment) on Icon component